+++ doc/authoring_command_modules/example_module_template/HISTORY.rst
@@ -3,7 +3,7 @@
 Release History
 ===============
 
-0.1.0
+0.0.1
 +++++
 
 * Initial release of module.