+++ src/command_modules/azure-cli-acs/HISTORY.rst
@@ -3,9 +3,13 @@
 Release History
 ===============
 
+2.0.25
+++++++
+* clarify `--disable-browser` argument
+
 2.0.24
 ++++++
-* [AKS] fix get-credentials command
+* fix get-credentials command
 * aks doesn't need to add role for SPN now
 
 2.0.23